modules/SceGxm fix compiler warnings

+++ vita3k/modules/SceGxm/SceGxm.cpp
@@ -711,7 +711,7 @@ EXPORT(int, sceGxmDraw, SceGxmContext *context, SceGxmPrimitiveType primType, Sc
         const auto vertex_program = context->record.vertex_program.get(host.mem);
         const auto program = vertex_program->program.get(host.mem);
 
-        const size_t size = program->default_uniform_buffer_count * 4;
+        const size_t size = (size_t)program->default_uniform_buffer_count * 4;
         const size_t next_used = context->state.vertex_ring_buffer_used + size;
         assert(next_used <= context->state.params.vertexRingBufferMemSize);
         if (next_used > context->state.params.vertexRingBufferMemSize) {
@@ -725,7 +725,7 @@ EXPORT(int, sceGxmDraw, SceGxmContext *context, SceGxmPrimitiveType primType, Sc
         const auto fragment_program = context->record.fragment_program.get(host.mem);
         const auto program = fragment_program->program.get(host.mem);
 
-        const size_t size = program->default_uniform_buffer_count * 4;
+        const size_t size = (size_t)program->default_uniform_buffer_count * 4;
         const size_t next_used = context->state.fragment_ring_buffer_used + size;
         assert(next_used <= context->state.params.fragmentRingBufferMemSize);
         if (next_used > context->state.params.fragmentRingBufferMemSize) {
